Be explicit with choice of integer data types

+++ spyns/lib/components.hpp
@@ -1,6 +1,7 @@
 #ifndef SPYNS_LIB_COMPONENTS_HPP
 #define SPYNS_LIB_COMPONENTS_HPP
 
+#include <cstdint>
 #include <string>
 #include <vector>
 
@@ -8,13 +9,13 @@
 
 namespace components {
 using Names = std::vector<std::string>;
-using Sublattices = std::vector<int>;
-using NumberNeighbors = std::vector<int>;
-using MapNeighbors = std::vector<int>;
-using ListOfNeighbors = std::vector<int>;
-using MapBilinearInteractions = std::vector<int>;
+using Sublattices = std::vector<uint64_t>;
+using NumberNeighbors = std::vector<uint64_t>;
+using MapNeighbors = std::vector<uint64_t>;
+using ListOfNeighbors = std::vector<uint64_t>;
+using MapBilinearInteractions = std::vector<uint64_t>;
 using ListOfBilinearInteractions = std::vector<double>;
-using BinaryStates = std::vector<int>;
+using BinaryStates = std::vector<int8_t>;
 using Vector3States = std::vector<data::Vector3>;
 }   // namespace components
 #endif   // SPYNS_LIB_COMPONENTS_HPP
